fix: tweak rest/wait with inhibited regen
Background terminology: CMD_REST triggers a "rest delay" if there's
hp/mp/ancestor hp/barbs to rest off (details configurable), or a "wait
delay" (which lasts a fixed amount of time) otherwise.

Before this commit:
* if you do CMD_REST with an inhibiting monster in view that can't reach
  you, a wait delay is started and lasts the full wait time (unless a
  reachable monster shows up of course). For example, a monster across
  deep water with no (direct-ish) path to you.
* if an inhibiting monster that can't reach you enters view during
  either a rest or wait delay, it counts as safe and nothing happens, so
  if the monster stays in view, a rest proceeds infinitely (which gets
  detected and triggers a crash). A wait finishes normally.

After this commit:
* CMD_REST doesn't work at all with an inhibiting monster in view
* both rest and wait delays are interrupted by an inhibiting monster,
  even if it can't reach you.

We'll see whether this is annoying for some reason. A more nuanced
behavior would be possible if there are common use-cases for waiting
with an inhibiting monster that can't reach you in view.

(It may still be possible to force an infinite rest by disabling seen
monster interrupts. One complication here is that there is currently no
particularly good way to detect when the -Regen status actually turns
on, because it all happens indirectly. But possibly the game should also
detect this and send a full_hp interrupt in this case.)