fix: Tweak monster placement in an early vault (dilly)
For `dilly_guarded_unrand_morg`, the second-highest tier monster glyph
probably shouldn't be an out-of-depth necrophage on D:2-D:3 when we're
also placing either Natasha or phantom, both of which are pretty strong
monsters at that depth. This vault is small and already places plenty of
additional monsters. Have the second-highest tier monster simply be
another derived undead on D:2-D:4, using it for something stronger only
when the vault is placed at higher depth.