docs: document stair delays (catern)
Previously, the stair-dancing penalty on re-entering levels was not
documented except insofar as it was possible to notice it from the time
indicator in-game. There's some discussion in the dev channel about
whether this delay "matters", but if it exists, it should be documented
at least minimally. Add a note to the manual and to the hints mode stair
description.