Remove scrolls of holy word
Scrolls of holy word are more or less the least interesting type of
scroll currently in the game. They're irrelevant for the vast majority
of the game, and when they are useful later on, it's not in a particularly
interesting way. They don't create exciting situations or memorable
moments, just inventory clutter. Let's cut em.

TSO's holy word remains in place... for now.