fix: Adjustments to an arrival vault
For `spicycebolla_arrival_teleportal_tutorial`, rename this vault to
`spicycebolla_arrival_transporter_tutorial`, since the feature in
question is called a transporter. Add a missing level connector to the
path leading from the hatch. Remove some SUBST and SHUFFLE, relying in
the default undefined glyph mapping to floor as well as the natural
vault rotations that occur for variation. Reglyph and reorder
transporter statements to straightforward entry/dest order, so the
transporter definitions are easier to read.