Tweak Gauntlet description (jumble)
The last line was intended to describe the path through the Gauntlet
arenas, but came across as saying 'once you enter a Gauntlet, the
only way out may be through the arenas & the minotaur'. You can see
the path through the arenas once you're there, so there's no need for
this text.