fix: unbreak UNIQ_LOST_IN_ABYSS
A bit too late, since it's now exceptionally rare for this to come in to
play, but oh well. Since 850bfe61e39c27fcd unique item status would get
(incorrectly, I believe) cast down to bool when setting
UNIQ_LOST_IN_ABYSS, and get converted to UNIQ_EXISTS as a consequence.
This tries to clarify the current statue of the API and restores what's
left of this mechanism.