fix: don't let help mark unrands as existing
e9b396fb1906e62 generates the unrand to show its info in help lookup,
and a side effect of the generation call marks the unrand as existing in
you.unique_items. In consequence, a player can force arbitrary unrands
that haven't yet generated to be (potentially) converted to fallback
randarts by looking them up. I'm not sure if this is good for players to
do or not but it's at least a very good bug.

Resolves: https://crawl.develz.org/mantis/view.php?id=12702