fix: Adjustments to an arrival vault (personjerry)
The vault erik_arrival_cave_to_temple had an area in a subvault that
could place monsters and loot fully encased in stone. Have this area
always place a door, introducing some clear stone so it won't be as much
of a spoiler and autoexplore trap.

The other subvault for this map placed monsters immediately behind a
door followed by nasty monsters and altars behind digable walls. Remove
the door to improve spoilers/autoexplore, and put the nasty area behind
a runed door. Tone down the Abyss monster set to not place demonic
crawlers on D:1, instead placing only the 5 class demon that can
generate in the abyss. Fix `no_item_gen` masking and also add
`no_monster_gen` masking to this section.