Re-rework animate dead
Make the zombies unlimited duration, but have them crumble when the
caster re-casts this spell. This is intended to avoid the boring play
pattern of repeatedly casting the spell at the start of every fight,
and to encourage somewhat more dynamic and fun decision-making.

Also reduce power scaling to make the spell more usable by 'hybrids'.