refactor: clean up handling of -wizard and -explore
Rather than having the CLO parsing code set an option value which then
interacts with the defaults in a complicated way, have fixup code check
the options directly after all option handling is done, and promote
WIZ_NEVER to WIZ_NO as needed. This has the benefit of getting a more
sensible interaction of the CLO and rc options (not that I would imagine
this comes up much); but the main goal is just to simplify the defaults
a bit.