refactor: standardize parsing of wiz_mode, explore_mode
This moves these options to be `MultipleChoiceGameOption`s for
standardized parsing, default setting, etc. The implementation in builds
that disable these is slightly different than previously. In DGL builds,
these options cannot be set but do produce an error (previously silent)
indicating that they are disabled for the build. CLO behavior is
unchanged. In !WIZARD builds, the option field is now present in the
Options object, but cannot be set via rc or command line so is basically
inert, aside from erroring (previously this case was also silent).

If the error is annoying for some reason it would be possible to restore
the old silent behavior, but I'm not seeing a lot of reasons why this
error would be bad.