fix: repair an ancient fixup
In current dcss, NUM_BOOKS drastically exceeds MAX_SUBTYPES. This
doesn't matter, because nothing (after this commit) assumes that books
are relevant to MAX_SUBTYPES, and this is why (I'm assuming) there was
no prior COMPILE_CHECK for this case. However, this one fixup assumed
that there were enough force_autopickup slots that there was a value for
NUM_BOOKS, leading to a crash on saves from this era. It's possible that
MAX_SUBTYPES should be increased to prevent future confusion, but I
instead just added a COMPILE_CHECK to validate the specific assumptions
current force_autopickup needs, and changed the fixup.

Found via the save in: https://crawl.develz.org/mantis/view.php?id=6308
(though this isn't either of the crashes noted there)