fix: rewrite webtiles menu pgup/pgdn
This was previously a bit of a mess; the behavior of the two was almost
never symmetric, and there were all sorts of messy interactions with
headers (because they are rendered in a smaller font). This rewrite
doesn't solve everything but it generally has much better behavior on
both of these points.

* refactor so that these are largely using shared code, making symmetry
  much easier
* redo pgup to use set_bottom_visible rather than do an ad-hoc size
  calculation (that was also wrong). This makes it much closer to how
  pgdn is implemented.
* when recalculating hover after paging, ignore headers. This overall
  leads to much more consistent behavior. (Though it still has some
  weird cases..) N.b. possibly this should be ported to local tiles too,
  but in practic local tiles paging (which uses a different approach) is
  less susceptible to these issues. Alternatively, webtiles paging could
  get the local tiles algorithm, which scrolls by pixels?
* heuristics for hover at the end

Also fix a bug for menu.last_visible calculation with size 1 menus.