refactor: convert `pregen_dungeon` to enum, clean up parsing
This option was previously implemented with the double bool for three
values strat; this converts it to a 3-valued enum and uses this to clean
up the parsing code as well. Sort of annoying to have yet another tiny
little enum.h around, but I guess that's what we do.