feat: add some basic monster depth info to objstat
Two new fields derived from mon-pick-data.h. These are a bit rough but
still informative in my testing.

RelativeDepth: the current depth minus the average depth based on
regular placement tables. So more negative is more out of depth, more
positive is more under depth. So for example a fire dragon placing on
D:9 gets RelativeDepth -11, because its average placement (rounded) is
20. Out-of-branch monsters get 0 for this, so the following Num stat is
what is used to calculate the average.

NumInBranch: Counts monsters that are generated by the regular branch
placement tables. For individual monsters, this will either be the total
count or 0.  (Maybe NumOutOfBranch would be more useful as a directly
viewable stat?) A few caveats: neither uniques nor derived undead ever
count as in-branch.

I did it like this because there is already an average function, but it
might also or instead be nice to have an explicit OoD number, maybe
something like the closest depth range edge minus current depth. E.g.
the fire dragon in the example above is also over-OoD, because its
minimum placement is D:14 (placed by a vault), and this is a bit hard to
directly reconstruct. Maybe later.