feat: add some monster OOD info to objstat
The relative depth values correlate with this, but it's interesting and
potentially useful to pull out OOD info specifically. At the moment for
large iters, the NumOOD field very rarely exceeds 0.01 for individual
monsters that do place in my testing, so I may need to tweak this
further for it to be useful.