fix: tweak precision for NumOOD in objstat
These are reporting rare enough events that with just to digits of
precision, a lot of stuff that is non-0 shows up either as 0.00 or 0.01.
An extra digit on top of the default seems to be enough that most values
can be compared when working in the 1000 iters range.

(cherry picked from commit 53bdb1016df8621e64d8913573d5fdc831def003)