New unrand: the trishula "Barbs of Condemnation"
The weapon of a long-gone archangel of judgment, this is a +8 trishula
that inflicts 4-8 turns of Anguish on hit. (That is, victims have their
damage mirrored back to them.) It also gives flying, rN+, and makes angel
noises. Fun!

It's a one-hander so that the anguish damage is more relevant (rather than
the weapon damage more quickly & directly murdering everything), a polearm
to synergize with allies, and a trishula for thematic & mechanical variety.
(We already have a perfectly good demon trident unrand.)

For more details, please read the description of Barachi in the Crawl manual.