Simplify Death Channel (Monkooky)
Remove the ability to recast while the duration is active, forcing
all specters to expire before you can get a fresh batch. This should
hopefully make gameplay with it a little more dynamic, and also
improve UI a bit - you no longer need to remember to recasting when
the expiration message appears.

It still seemed quite strong in testing, but other adjustments can
always be made if need be.