Revert "Revert "Make new stairs fast after crashing (advil)""
There's a bug here, which is that the game state is saved before stair
travel time (as well as seen_monsters_react) is applied, but this
happens independently of this flag, and so this revert did not fix it.
So killing the crawl process, forcing a crash, immediately after stair
descent always rolls back whatever stair time there was as well as the
initial hits.  Also possibly a bug that the crash isn't saving? I'm not
sure why it wouldn't.

Also, I kind of want to say that force-quitting a local game is
essentially a form of save-scumming and so I'm slightly mixed on whether
this is really what I'd call a bug.

This reverts commit f3dacdeeaaba24e7937b459bf04e6c4ca36dcd6f.