fix: save after world_reacts on level entry
There's a save call in `floor_transition`, and a new level is saved on
generation. However, these save states are not at a point where the
player can ever have control over the game, because there's always a
following world_reacts cycle. Because of this, if a crash happens after
these saves but before a real save, the game is rolled back to a time
that is impossible to get to in any other way -- time 0 after level
entry. This is the bug discovered by personjerry and mentioned in
f3dacdeeaaba2. To be clear, it's *not* (imo) a bug that you can roll
back time by intentionally force quitting or killing a local game in
this way, that's just a variant of ordinary savescumming -- this commit
doesn't prevent using force quits to roll back time[1]. But what it does
ensure is that a crash rolls back time to a possible game state. The
implementation is very brute force and adds Yet Another Bool to
`crawl_state`. Along the way I fixed what I think is a bug in sprint
save checkpoints -- the level state wouldn't get saved, so a crash in
sprint could at least in principle lead to a save with a very
inconsistent player / level state.

As a side note, the fact that outside of regular save+quit the player
and the level are typically saved independently (something I hadn't
grasped before this fix), could lead to all sorts of other weirdnesses
on crash.

[1] This could be desireable but it's not entirely clear to me how to do
it, and it's also quite platform-specific.