Move orbs to lightgrey, animal skins to red.
Orbs had the same colour as cloaks and scarves, despite being in a
completely different slot.  Rather than move orbs to cyan with their
slotmates, put them on a mostly-free colour.

This requires, or at least suggests, moving the one other lightgrey
armour, animal skins, to another colour.  Since skins are functionally
identical to robes (this is, as far as I know, the only remaining
instance of functionally-identical items distinguished purely for
flavour reasons), put the two on the same colour.

Also move CPA in the relevant switch, to keep the (now fewer) white
armours together.