refactor: remove mons_zombifiable
88678701b151 made this equivalent (indirectly) to
mons_class_can_be_zombified, and I'm now pretty sure this is correct.
The only reason that has come up to have them distinct is to allow some
kinds of monsters to place in vaults but not generate randomly, but this
can also be accomplished by setting the M_NO_GEN_DERIVED flag, so the
only function of both of these existing at this point is confusion.

Also removes a bunch of lua stuff that I think is now useless since
corpses are purely cosmetic; if I'm wrong it's easy enough to revert.
Possibly what is needed is now similar functions that act on monsters?