fix: handling of animating objects on banishment
This reverts an incidental part of df715e3894bdc; after this commit the
animating object for dancing weapons was getting unconditionally
destroyed on banishment. The ENCH_ABJ check here is necessary to
identify animated objects that are temporary but not summoned. In
addition, I think it is correct to only set UNIQ_LOST_IN_ABYSS on
temporarily animated objects, because otherwise (in principle)
duplicates would be possible.

I'm not 100% sure that KILL_BANISHED will do something sensible on a
temp animated weapon, independent of this codepath though? It may still
go on the transit list...