fix: don't wipe out Ru muts with change_species_to
This code was promoted from a wizmode function, where things didn't need
to be absolutely perfect, to being used in fixups. Among other issues,
it overlooked Ru sacs, which are implemented as innate mutations. I'm
not perfectly confident that the following code won't mess with sacs in
some way I don't understand, but this fix works for the straightforward
cases.