fix: move a bunch of option defaults out of dat/defaults
I would prefer if only lua options had their defaults set in these
files. We are a ways from there but these two really don't need to be
done in an include; one of them even had its default split between c++
and the include. Everything that is left here is either lua or a
text_pattern (where for reasons I haven't discerned, text_pattern
defaults never seem to be set in c++).