Turn Sandblast into Slowblast
The idea of a spell that uses up ammunition of some kind is pretty cool.
You can use less powerful spells to conserve ammo, and then use the big
spell for special occasions - if you have other spells available, there's
no inconvenience similar to switching between different weapons.

However, at level 1, you won't *have* other spells available. So the only
way to preserve stones for Sandblast is to switch to melee for chaff...
which isn't very exciting!

Instead, let's play with a new mechanic that I've wanted to try out for
Earth Magic for a while. Sandblast now no longer consumes stones, but
instead takes twice as long to cast as normal spells. In testing, this
gave it a distinctive and spicy character, but still felt quite powerful.
(It's probably a slight nerf to D:1 EE and maybe a buff for D:2-3, which
seems fine.)

I'll keep an eye out for other places to use the ammo mechanic, though it
might not involve stones.