fix: don't let broodmother spawning crash the arena
The foes for summon spiders were set wrong. This commit generalizes some
code and should perhaps be applied further for the sake of
future-proofing, since any of these summoning spells that just sets the
foe to MHITYOU would crash in the arena if it were ever callable in the
arena.