New item: wand of light
Intended as an alternate for wands of acid, wands of light fire a beam
that blinds targets for a duration dependent on HD. The beam checks AC
and EV normally, but no resists. It's otherwise identical to acid.

If it hits a player, it has a 1/3 chance of confusing them for 2-3 turns.
(Since players can't be blinded.)

The goal is that this has a broadly similar use case to acid (high damage
beam that ignores most resists & applies a debilitating effect), but plays
somewhat differently. It'll be more useful for stabbers, and somewhat helpful
defensively (by reducing enemy accuracy instead of AC).

I'm very undecided on the blinding effect duration.

Buff wand of light