refactor: updates to webtiles keyhandling
This is my latest take on how to do webtiles keyhandling in a way that
points towards a more modern approach, while supporting old game
versions. Followup to 7b1be97da7.

* For versions following this commit, convert numpad and function key
  handling to use modern `code` values rather than keycodes. This has
  widespread browser support in 2022 and should be more robust than the
  previous keycode-based approach.
* Remove keycode overrides from game.js, with a new division of labor
  and clearer handling for legacy keycodes. Now all the client needs to
  do to use updated keyhandling is enable `code` mappings.
* Support for some new internal keycodes: ctrl+shift bindings; modified
  space, del, enter, bksp, esc; numpad =/enter, F keys through F19, fix
  delete handling (this is the only one that works on older versions).

Still to be done: fix alt handling, convert modifier handling to use
`code`.

This commit attempts to support all versions and all possible cache
states for non-versioned code (which on some servers may require a
server restart to refresh). We'll see if that actually works.