feat: Rename the Hedge Wizard abbreviation to HW
When Wizard was renamed to Hedge Wizard in 0.26, we avoided changing the
background abbreviation. This was to avoid the hassle of retroactively
renaming combos, but it left us in an awkward situation where the two
class names were different yet the combo names remained unchanged. We're
sure we want to regard the historical Wizard class as equivalent to
current Hedge Wizard, and with release coming soon, now is a good time
to rename the abbreviation and do a retroactive rename in Sequell.

This commit renames the abbreviation Wz to HW and renames all related
enums and image filenames for good measure. I've also updated the lookup
key for background descriptions in all language files that previously
had entries for Wizard, but I've not changed any of the language text.