fix: float issue for webtiles pgdn / last menu item calculations
After the new pgdn code from c0181ea820b1, in some cases the last menu
item was not being calculated correctly because of minute (~ 0.1)
overlap of the last menu item with the footer, I assume because of
floating point inaccuracy. This caused pagedown to not consistently move
to the last element in a menu that is larger than one page. This commit
tries to fix this by just adding a bit of heuristic margin; I don't
think this error should ever be more than 1px or so.