feat: explore_wall_bias tweaks
* don't consider just cardinal directions in checking for adjacent
  walls.  (Fair warning, I don't know why it originally did this, so
  this change could be breaking something?)
* let explore_wall_bias be negative, to prioritize open space.

I should add that the use case of this option remains somewhat
theoretical. There seems to be some historical idea that setting this to
positive values would speed up eploration, but in my testing this is
really far from correct: it adds sometimes 100s of turns to exploring a
level, because it can drastically increase the amount of backtracking.
Similar for negative values. It *might* be useful for meteorans as it
possibly deprioritizes things they don't want to explore anyways, but
the total explore time if you do a full explore is definitely still
increased. In my testing for negative values what tends to get
deprioritized are corners and corridors, which is what I was hoping for.
But it's hard to say anything conclusive about whether this would be an
improvement..