feat: add zot clock to dungeon overview
This displays in both the dungeon overview popup, and in morgues.
Whether it displays follows the usual rules for showing the zot clock:
it is forced by always_show_zot or meteorae, and otherwise only shows
info if the bezotting level is >0. This also clearly indicates when the
player is in death range as well as which branches they need to watch
out for. Unfortunately, though I tried very hard, I just could not get
this info into a three-column display. (I tried both inline approaches,
and two-row branch displays; the former doesn't have enough space, and
the latter looked quite bad.) Therefore, if any zot clocks are shown, it
flips to a two-column display. Possibly confusing?

I should also note that there are enough special cases here that I
really doubt I got all the spacing to look right in all cases (as much
as I did try to generalize the code for this a bit).