feat: indicate zot clock death risk in status light
This might be overkill, since I think the likelihood of a player drained
this much actually dying of the zot clock itself is rather low, but it
seems useful to try out. In the end, rather than fancy punctuation, I
simply added ", death" if the next zot clock 0 would kill.  The color
tweaks here bring this in line with the overview screen, though in
contrast to there, the color is always stepped up if the player is
drained enough to be risking death. (If `always_show_zot` is off though,
for non-meteorae, clocks >=1000 turns are still not shown, though.)

I also tweaked magenta->lightmagenta mainly for the sake of webtiles
where regular magenta looks a bit dull (possibly just my badly
calibrated monitor).