vaults: new demonic serial vault
Encountering "common demons" at a time when they're a threat and not
summons is a rare occurance in crawl. This serial vault, a shameless
ripoff of serial_undead, places some demonic themed decor on a dungeon
level to give it a "demonic feel". Spooky.

Featuring demonic trees and plants, plus 5, 4, and 3 tier demons.