feat: abyssal rune detection
In newabyss players want to get in and out of the abyss as quickly as
possible, but there's a lot of variation in how quickly people find the
run, and it can really drag on with no rewards if this doesn't happen
quickly. When you're at a rune depth, there's not a lot you can do to
find the rune beyond just wandering and/or looking for vaults via
giveaway tiles (these could be removed if this current commit goes
forward).  This commit implements an idea where all players detect the
presence and map location of the abyssal rune, and also detect if the
map shifts and it vanishes. This gives players a much more concrete
goal, as long as the rune has generated.

I am still not entirely sure if this is a good idea, but in my testing
it works way better than I expected. The rune typically generates pretty
far from you and then you have a bit of a race towards it, both (a)
simply navigating the abyss between you and it, and (b) figuring out how
to even get into the vault. I'd theorycrafted this as making it too easy
to find the rune, but I'm not sure it (typically) is at all. But, it
probably needs more playtesting.

Also, by the nature of abyss code, I am at best medium sure that I added
this hook in all the places that it needs to be in.