fix: rewrite abyssal rune detection
The first pass implementation relied on being able to store rune memory
purely in map_knowledge, and use this at various points in time to
detect both rune generation and removal. The abyss code makes this
impractical. The rewritten version stores rune memory in a player prop
that gets synced with map_knowledge; this is much more robust in my
testing.