fix: adjust floor_crystal_squares in abyss rune vaults
In my testing, I'm not sure this tile needs to be abolished. The
previous issue was that it was (a) spoilery, and (b) spoilery in a way
that console players couldn't access. However, with abyss rune
detection, it is generally obvious when you are nearing the rune vault
(even to players who don't know exactly what a vault is, I think it's
obvious that there is something special nearby). This tile (as well as
th associated coloring, and the empty arches) are decent flavor in that
case. So this commit takes a pretty light hand: the cases in my testing
where it still seems to run up against a-b are cases where a reasonably
sized rune vault has only one entrance, but the vault used this tile
sprinkled around the edge to indicate the vault boundary from sides
where it wasn't otherwise apparent. I have removed it from mostly just
these cases. (And possibly haven't gotten this especially perfect...)

It also seems to me like it might be better design not to have such rune
vaults in the abyss, either relying more on multiple entrances, or on
clear tiles to cue the player (Similar to what the transporter vaults
tend to do.) The problem is that in combination with abyss terrain
generation, these vaults can often be quite frustrating to access even
if you know they are there.