feat: make -Scroll and -Potion regular 2-tier bad mutations
Two tiers are added to both MUT_READ_SAFETY and MUT_DRINK_SAFETY
mutations to make it more in line with the other bad mutations, and the
latter is added to the general bad mutation pool.

To make the 2 tiers work correctly the _is_damage_threatening function
has an added variable to decide on the safety level with which to
evaluate the threat. Ru's sacrifice now sets the mutation at tier 2,
which is the same severity as before this commit. This commit also
adjusts the weight of both mutations to 7 (average) from 10 (most common).

[ Committer's note: Squashed, rebased, checked whitespace. Closes #2573 ]