Add 3 tiers to the MUT_READ_SAFETY mutation
The 3 tiers are added to the MUT_READ_SAFETY mutation to make it more in line with the other bad mutations. To make the 3 tiers work correctly the _is_damage_threatening function has an added variable to decide on the safety level with which to evaluate the threat.  Ru's sacrifice now sets the mutation at tier 3, which is the same severity as before this commit. This commit also adjusts the weight of the mutation to 7(average) from 10(most common) and adds the MUT_ACUTE_VISION as a conflict mutation.