feat: re-vitalize Vitality
Regeneration and 15 hit points aren't much. Vitality has been on the "to
redesign" list for a long time. Some of the possibilities discussed are
very close to what ultimately became the dreamshard amulet. It was also
suggested, and discarded, as a possible early-game unrand.

Let's buff it out of this awkward place. The 15 hp are traded for two
pips of MP regen and an additional hp regen pip. Lore has it the amulet
kept Borgnjor alive in her final days before descending into undeath, so
it's also now evil.