feat: reduce stone availability and make them always mulch
Stones are very good on D:1-2, and while they quickly fall off remain
useful for characters for a bit longer, or as a "poke the monster
quietly before walking away to pull a small number of foes" tool.

They're also all over the dungeon in massive numbers, and are re-usable
to boot. Now that they're not used for Sandblast, their primary gameplay
function is early game plinking. To make this plinking a bit more
interesting this commit reduces the supply of stones in each pile, and
makes stones always mulch when used. In the very early game this should
put significant pressure on the use of stones in an interesting way,
with them becoming plentiful closer to the time the player has better
things to do than plink away.