revert: unrewarding abyss
This reverts most of the "no-XP" abyss changes, specifically:

fcf54068 feat: prevent gaining experience in the Abyss
d764ef0d feat: give an XP reward for the abyssal rune
010f2a22 feat: don't grant piety in the Abyss
0647d936 Refactor piety gain prevention in the Abyss
6c1bee59 feat: reduce wretched star malmutation rate
050a5067 fix: turn down wretched stars corrupting pulse a bit more

The no XP abyss did stop abyss scumming, but at too great a cost. While
some players enjoyed the stealth-emphasis-increase of the XP being
concentrated in the rune, the overall result wasn't good.

So this reverts the changes that revolve around XP, in preparation for
the following commits.

Not reverted are:
- the changes to exit and stair spawn rates. These remain increased, in
  anticipation of the subsequent anti-scum mechanic making them
  desireable.
- the removal of items outside of vaults. Items still have the potential
  to be a grind-trap (more a trap than "optimal" scumming), but the
  abyss isn't supposed to be a place players want to linger so let's not
  give them too many reasons to stay beyond a rune.