feat: force a minimum entry depth on the Abyss
In lieu of the previous anti-scumming mechanism, which threw the baby
out with the bathwater and created an unpleasant experience with a lot
of special cases, this introduces a minimum depth mechanic for the
abyss.

Upon entry from any source (distortion, banishment, exiting pandemonium
through the abyss, entering an abyss portal) the player is sent to the
highest depth so far achieved.

In combination with the rune-guide improvements, a continued dearth
of items, and the XL scaled depth-increasing-on-teleport mechanic, this
should give a better experience to the recently punished mid-game
banishment victim and the 3rd rune hunter, while still dissuading the
scummer (who will rapidly find themselves on A:5) than the no-xp abyss
did. If A:5 is not harsh enough to be a deterrant we can add greater
depths as needed.