feat: select the size of the Temple in C++
Temple size has a major impact on god selection: a big Temple means
getting your choice god-but it might take until D:7. A small temple
scatters the gods over the overflow range. Both experiences are good for
the game, but before this commit the chance of a given temple size was
controlled by the sundry weights of the different temple maps. This
control was brittle, and indeed b444dcf96dcc4 marked a big shift in the
prevalence of larger temples (thanks to acrobat for pointing out how
large this impact was), that was not intended.

This commit changes temple selection to work as follows:

First a 1% roll decides to select a rare temple. Current rare temples
are the Lugonu, Jiyva, and Ignis temples. Rare temples are those that
should create special moments outside of the usual temple system.

Failing the 1% roll, the temple size is rolled. Currently this is 9 +
random2avg(13,2), but this can be adjusted (and will be to accomodate
some smaller temples coming in subsequent commits).