Allow banished monsters to appear after Abyss:1 (acrobat)
Previously, there was an oversight (?) in which banished monsters
would only ever appear on Abyss:1. Make them follow you through the
whole Abyss instead. This is a little silly (are they taking the
stairs down?!), but prevents you from losing the Royal Jelly on a
lower Abyss level, for example.