fix: don't use levelgen checks for TRJ splitting (particleface)
The TRJ splitting code doesn't use normal summon placement code (which
would typically be `find_habitable_spot_near`), but rather repurposes
some levelgen placement code, that has the right adjacenty check for
TRJ.  However, because this code is otherwise used only in levelgen, it
enforced an `MMT_NO_MONS` mask check. A result is that if you can manage
to fight TRJ in the loot vaults, it won't split.

This commit simply conditionalizes the levelgen check, so that the
placement-finding code can be used either way. On a quick glance, this
should bring it in line with regular summon placement; there isn't an
explicit los check but I expect that's approximated by the floodfill.