fix: reorder skills on the `m` screen
The skill order from 586177b653 does make sense:
* Throwing no longer crosstrains with Slings (now Ranged Weapons), so
  they don't have to be in the same group;
* both Throwing and Evocations rely on consumables.

On the other hand, Throwing and Invocations have nothing in common and
should not be in the same group. It's better to have Throwing with all
the other Dex-based skills, right below the Str-based skills. Also,
this allows to have all the physical/non-magic skills on the left, and
the magic/divine skills on the right. Evocations also belongs there,
because any sufficiently advanced technology is indistinguishable from
magic :).