feat: reduce the +stab effect on Spriggan's Knife by 50%
Most monsters die in a single hit from the Spriggan's Knife; those that survive take massive damage and can be quickly dispatched with a couple more stabs. Stabs that result from netted/webbed/petrifying/invisible/confused often leave the target open for further stabbing, as the condition still applies. This is especially overpowered with irresistible status effects such as that from Line Pass.

To make this less of a guaranteed instakill, add a coinflip(). It may take a few tries before the target dies.