feat: enable small Temples
With the addition of small Temple maps, lower the minimum Temple size to 6.

The distribution used has a picewise linear probability density
function, starting with a 2.5% chance of a temple of size 6, rising to
its maximum at size 13.5, (roughly 10% each for sizes 13 and 14) and
then decreasing to a 2.5% chance of a temple of size 21. This is a bit
more verbose than using 6 + random2avg(16,2), but the latter makes the
extremes too small.

A distribution that falls off harder in the direction of larger temples
might be interesting but it's a somewhat tricky balancing act.