Slash Palentonga roll attack noise
When I designed Palentongas, I added 12 noise to them so that it wouldn't
always be correct to use roll attacks. Now that we've had a while to play
with them, I suspect this was the wrong call. Roll attacks are Palentonga's
cool thing - players should want to use them very heavily, and we should
design the UI to support that.

Reduce noise to one-third, 4 noise. This is mostly there as a justification
for preventing roll-sleep-stabs, but isn't really noisier than the melee
combat itself.