Rename Palentonga to Armataur
The name 'palentonga' was intended to be a bit more subtle than 'felid',
'octopode' or even 'barachian' - to be an 'animal-person' species name that
merely hinted toward the species (pangolin), rather than shouting it from
the rooftops. But of course, they weren't exactly pangolins - they were giant,
and had four legs plus two arms...

This was too many things at once! If the concept was simple and familiar, we
could have used a more exotic name, and the rest of the flavour would have
put players on the right path. But as it was, players were lost. There was no
resonance.

So, let's rename em. 'Armataur' is a simple 'armadillo' + 'centaur' combination.
(Yes, I know that properly they should be 'cenadillos'...) The abbreviation is AT,
to avoid ambiguity with the Artificer start. It's exciting, it's hip, it's fun.
Let's try it out.