docs: comment Menu class hierarchy
I'm tired of trying to keep all this in my head and inevitably
forgetting StackFiveMenu or whatever. No doubt this comment will go
stale at some point, and there are ways of programmatically generating
this, but I think it's still better to have as just a text comment in an
easy place than not.

Also, remove a reference to a nonexistent class.