feat: add CMD_MENU_EXAMINE and support code
The Menu superclass had some support for the concept of describing a
menu item, via ACT_EXAMINE and all the cycling stuff, and nearly every
Menu subclass uses it somehow, but for whatever reason handling was left
to manual subclass code.  This commit generalizes the standard pattern
into the superclass, hopefully making it more straightforward to write
code that does whatever description popup is needed. Along the way, this
change makes it possible to add a distinct "examine selected menu item"
command, which I have bound to `\\` and `'`. The latter replaces a
rather ancient compatibility binding, we'll see if this bothers people.

Some misc refactoring that happened as part of this:
* clean up acquirement menu a bit
* enable hover in mutation menu, various deck menus
* improve some shop menu behaviors, stash search menu behaviors
* fix a weird casting preselect bug