Slow fuzzing (Chei, Trog, do_more failthrough):
The slow_player() function already implements checks to limit the maximum duration of slow, and they occur downstream of the checks in god-wrath.cc -- The checks in god-wrath can pass if DUR_SLOW is between 100 * BASELINE_DELAY and 180 * BASELINE_DELAY, but the checks in slow_player will fail and the player won't actually get slowed. I'm removing the checks in god-wrath.cc.