Lugonu wrath minor minion fuzzing
Lugonu's wrath has a major case and a minor case, if it rolls major the number of lesser servants is random and there's a guaranteed major (and maybe a bonus major). If it rolls minor, the number of lesser servants is not random. I've replaced the flat 1 + XL/7 with random2(3) + XL/7, this gives a 1/3 chance that a low-XL minor roll will return no monsters which stinks, so I've special-cased it: you're guaranteed one.