Zin Wrath: Oops, all badmuts!
Zin's wrath has a 33% chance to do literally nothing if you have mutations but all of them are bad. This is silly: Zin hates chaotic creatures, so you should be punished for having bad mutations, not rewarded! Zin should recite at you rather than do nothing if you're so chaotic that you're beyond saving. So, I've made it do that. Making the remove mutations function a bool allows us to check if it fails (and it already did that for the purpose of generating the right message), and reordering the cases to generate an intentional fall-through in the switch statement.