fix: various issues with the `F` menu
Previous commits broke some of the keybinds for this menu, which needed
attention anyways. The "throw any item" menu was different from the *
menu only if you had a wielded launcher at this point, so it was a
little silly to have both. Also, fix a bunch of the messaging,
generalize some of the validity checks.