New unrand: toga "Victory"
This is an early unrand with a variable bonus based on kills:

As the player kills threatening monsters, this robe gradually gains AC, Int
and Slay, capping out at +7 each.

Beware -- Quaffing potions or reading scrolls in combat (with monsters
visible) resets these bonuses to zero!

[Committer's note: squashed commits and updated commit message. Fixes some
logic for when to apply the bonuses, and adjusts messages/description.
Closes #2920. -kate]